+++ packages/hot-updater/tsdown.config.ts
@@ -4,11 +4,21 @@ export default defineConfig([
   {
     entry: {
       index: "./src/index.ts",
-      config: "./src/config.ts",
     },
-    format: ["esm", "cjs"],
+    format: ["esm"],
     outDir: "dist",
     dts: true,
+    failOnWarn: true,
+    shims: true,
+  },
+  {
+    entry: {
+      config: "./src/config.ts",
+    },
+    format: ["esm"],
+    outDir: "dist",
+    dts: true,
+    clean: false,
     failOnWarn: true,
     shims: true,
   },