+++ packages/console/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @hot-updater/console
 
+## 0.30.7
+
+### Patch Changes
+
+- Updated dependencies [03fd179]
+  - @hot-updater/cli-tools@0.30.7
+
 ## 0.30.6
 
 ### Patch Changes