+++ plugins/expo/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @hot-updater/expo
 
+## 0.30.7
+
+### Patch Changes
+
+- Updated dependencies [03fd179]
+  - @hot-updater/cli-tools@0.30.7
+  - @hot-updater/bare@0.30.7
+  - @hot-updater/plugin-core@0.30.7
+
 ## 0.30.6
 
 ### Patch Changes