+++ plugins/js/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @hot-updater/js
 
+## 0.30.7
+
+### Patch Changes
+
+- @hot-updater/core@0.30.7
+
 ## 0.30.6
 
 ### Patch Changes